Convert WPOrderReviewLineItems to TypeScript
